% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{parseGame}
\alias{parseGame}
\title{Parse an enhanced game json file into tibbles suitable for summarization}
\usage{
parseGame(enhancedGameJSON, masterRoster = NULL)
}
\arguments{
\item{enhancedGameJSON}{path to JSON file containing enhanced play information}

\item{masterRoster}{data frame containing PlayerID and attributes of each player (generally
created from a retrosheet .ROS file)}
}
\value{
A list containing five tibbles:
\describe{
  \item{BatterPlays}{One record for each play involving a batter}
  \item{Defense}{One record for each error, assist, and putout}
  \item{RunsScored}{One record for each run scored}
  \item{StolenBases}{One record for each stolen base}
  \item{CaughtStealing}{One record for each caught-stealing}
}
}
\description{
Note that the input json data must be in the \code{baseball-tools} enhanced format. See the
\href{https://github.com/scottcame/baseball-tools/tree/master/jslib}{baseball-tools Javascript library}
for more information and guidance on how to enhance raw retrosheet input data.
}
