% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{parseGames}
\alias{parseGames}
\title{Parse a collection of enhanced game files in a directory.}
\usage{
parseGames(sourceDir, filePattern = ".+-enhanced.json")
}
\arguments{
\item{sourceDir}{the directory from which to read source files}

\item{filePattern}{pattern defining files to read}
}
\description{
Parse a collection of enhanced game files in a directory.
}
