% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{summarizeGames}
\alias{summarizeGames}
\title{Produce summary stats from a list of parsed game data}
\usage{
summarizeGames(playDataList, teamID)
}
\arguments{
\item{playDataList}{list of parsed game data}

\item{teamID}{team for whom to produce statistics}
}
\description{
Produce summary stats from a list of parsed game data
}
\examples{
parseGames('~/my-scoresheet-data-dir') \%>\% summarizeGames()
}
